Fixes to actually run code
This is the version to run the first (trivial) infinite loop
correctly.  I hacked up the loader slightly to replace one of the
(nonexistent) pages with a static byte array.

z80 code executed:

$4000:
	JP	$4000

4000  c3 40 00
4003

This version also draws the opcode executed on the screen, to aid in
debugging the emulator:

        move.b  d0,$4c00+32*(128/8)

Also, it turns out that I was mixing up the emulated SP (a4) and emulated
PC (a6).  That has been fixed.

Further, it seems that movea defaults to a word operation in a68k.
This led to pointer corruption, which has been fixed.